[완료] 1. MonsterSpawnScheduleManager. 작업중. 라운드별 작업 중. 라운드 연동 완료.[80%]. 2. 모든 테스트 완료. 3. 실제 몬스터에 데미지 들어가는 확인 필요.
[버그]
0. 전에 없던 카드 셀렉터 창 시작시 에러 발생.
1.버그창에 에러 표시문제 해결
- 블랫 프리팹에 DoDestroy 스크립트 빠져서 생기는 에러

[참고]
0. 에어스트라이크 게이지 및 1제외 나머지 추가 필요함.
0. 플레이어 콜라이더와 바닥 콜라이더가 걸려서 자꾸 멈추는 느낌 있었음. 생성시 오브젝트 높이 고려해서 생성 시켜야함.
1. 플레이어 리포지션시 FollowCam 역시 같이 좌표로 이동 시켜야 함.
1. 데미지 UI의 경우 처음부터 맵 상에 존재하는 최대 몬스터 제한이 필요 할듯함. 느려짐 발생 가능성 있음. 레벨 디자인 하면서 조정이 필요.

[작업예정]
0. 재장선시 리로드 UI 메시지 데미지 창에 출력 -> 사운드 같이
0. 스테이지 정상호출 부가 작업 해야함. - 씬 이동시 플레이어 디스에이블.
0. 프랍 오브젝트 static 비활성화 시켜야함.
1. 몬스터 활성화 및 포지션 재조정.
2. 몬스터 죽었을때 데드 연출 변경필요.
3. 몬스터 생성시 에러 확인.

+++ Assets/WoosanStudio/Zombie Shooter/Scripts/Card/ICard.cs
@@ -73,7 +73,7 @@ namespace WoosanStudio.ZombieShooter
         /// <summary>
         /// 레벨 1업에 증가하는 Value수치
         /// </summary>
-        int IncreasedValuePerLevelUp { get; }
+        float IncreasedValuePerLevelUp { get; }
 
         /// <summary>
         /// 해당 프로퍼티 설명
